+++ test/text_search_test.dart
@@ -60,14 +60,18 @@ void main() {
   withPostgresServer('tsquery', (server) {
     test('read and re-read queries', () async {
       final queries = <String, String?>{
-        'x': null,
-        '!x': null,
-        'x & y': '(x & y)',
-        'x | y': '(x | y)',
-        'x <-> y': 'x <1> y',
-        'x <4> y': null,
-        'x & !(y <2> z)': '(x & !y <2> z)',
-        'x & y & z & zz': '(x & y & z & zz)',
+        'x': "'x'",
+        '!x': "!'x'",
+        'x & y': "('x' & 'y')",
+        'x | y': "('x' | 'y')",
+        'x <-> y': "'x' <1> 'y'",
+        'x <4> y': "'x' <4> 'y'",
+        'x & !(y <2> z)': "('x' & !'y' <2> 'z')",
+        'x & y & z & zz': "('x' & 'y' & 'z' & 'zz')",
+        'x:A': "'x':A",
+        'x:*': "'x':*",
+        'x:A*B': "'x':*AB",
+        'x:B & y:AC': "('x':B & 'y':AC)",
       };
       final c = await server.newConnection();
       for (final e in queries.entries) {
@@ -101,7 +105,7 @@ void main() {
 
       await expectMatch(
         vector,
-        TsQuery.lexeme('cde'),
+        TsQuery.lexeme('cd', prefix: true),
         true,
       );
       await expectMatch(